Volume.cc: reduce internal code duplication a bit
plenty of methods just choose a message format string and
then interpolate some values into that.

now we put the format strings into an std::map and select from
that based on inputs.

plus a little improved formatting in comments for translators.

most storage/*.cc files are ripe for this treatment.